+++ src/extension.ts
@@ -2,6 +2,15 @@ import * as vs from "vscode";
 import * as path from "path";
 
 import { Documenter } from "./documenter";
+import { TextDocument, Position, CancellationToken, CompletionItem, CompletionItemKind, Range } from "vscode";
+
+const languages = [
+    "javascript",
+    "typescript",
+    "vue",
+    "javascriptreact",
+    "typescriptreact"
+];
 
 let documenter: Documenter;
 
@@ -12,17 +21,13 @@ function lazyInitializeDocumenter() {
 }
 
 function languageIsSupported(document: vs.TextDocument) {
-    return (document.languageId === "javascript" ||
-        document.languageId === "typescript" ||
-        document.languageId === "vue" ||
-        document.languageId === "javascriptreact" ||
-        document.languageId === "typescriptreact" ||
+    return (languages.findIndex(l => document.languageId === l) !== -1 ||
         path.extname(document.fileName) === ".vue");
 }
 
 function verifyLanguageSupport(document: vs.TextDocument, commandName: string) {
     if (!languageIsSupported(document)) {
-        vs.window.showWarningMessage(`Sorry! '${commandName}' currently supports JavaScript and TypeScript only.`);
+        vs.window.showWarningMessage(`Sorry! '${commandName}' currently only supports JavaScript and TypeScript.`);
         return false;
     }
 
@@ -44,45 +49,51 @@ function runCommand(commandName: string, document: vs.TextDocument, implFunc: ()
     }
 }
 
+// Thanks, @mjbvz!
+class DocThisCompletionItem extends CompletionItem {
+    constructor(document: TextDocument, position: Position) {
+        super("/** Document This */", CompletionItemKind.Snippet);
+        this.insertText = "";
+        this.sortText = "\0";
+
+        const line = document.lineAt(position.line).text;
+        const prefix = line.slice(0, position.character).match(/\/\**\s*$/);
+        const suffix = line.slice(position.character).match(/^\s*\**\//);
+        const start = position.translate(0, prefix ? -prefix[0].length : 0);
+        this.range = new Range(
+            start,
+            position.translate(0, suffix ? suffix[0].length : 0));
+
+        this.command = {
+            title: "Document This",
+            command: "docthis.documentThis",
+            arguments: [true]
+        };
+    }
+}
+
 export function activate(context: vs.ExtensionContext): void {
-    context.subscriptions.push(vs.workspace.onDidChangeTextDocument(e => {
-        if (!vs.workspace.getConfiguration().get("docthis.automaticForBlockComments", true)) {
-            return;
-        }
+    context.subscriptions.push(vs.languages.registerCompletionItemProvider(
+        languages,
+        {
+            provideCompletionItems: (document: TextDocument, position: Position, token: CancellationToken) => {
+                const line = document.lineAt(position.line).text;
+                const prefix = line.slice(0, position.character);
 
-        if (!languageIsSupported(e.document)) {
-            return;
-        }
-
-        const editor = vs.window.activeTextEditor;
-        if (editor.document !== e.document) {
-            return;
-        }
-
-        if (e.contentChanges.length > 1) {
-            return;
-        }
-
-        const change = e.contentChanges[0];
-        if (change && change.text === "* */") {
-            lazyInitializeDocumenter();
-            setTimeout(() => {
-                try {
-                    documenter.automaticDocument(editor);
+                if (prefix.match(/^\s*$|\/\*\*\s*$|^\s*\/\*\*+\s*$/)) {
+                    return [new DocThisCompletionItem(document, position)];
                 }
-                catch (ex) {
-                    /** vs.window.showErrorMessage("docthis: Failed to document at current position: " + ex.message); */
-                    console.error("docthis: Failed to document at current position.");
-                }
-            }, 0);
-        }
-    }));
 
-    context.subscriptions.push(vs.commands.registerCommand("docthis.documentThis", () => {
+                return;
+            }
+        },
+        "/", "*"));
+
+    context.subscriptions.push(vs.commands.registerCommand("docthis.documentThis", (forCompletion: boolean) => {
         const commandName = "Document This";
 
         runCommand(commandName, vs.window.activeTextEditor.document, () => {
-            documenter.documentThis(vs.window.activeTextEditor, commandName);
+            documenter.documentThis(vs.window.activeTextEditor, commandName, forCompletion);
         });
     }));