+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 0.6.0
+- Now uses a completion provider instead of automatic insertion when typing '/**'.
+No longer conflicts with VSCode's built in JSDoc generation!
+
 ### 0.5.0
 - Support documentation generation for class methods.